etl/requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
minio==7.2.0
python-dotenv==1.0.0
schedule==1.2.0
pyspark==3.5.0
pyarrow==14.0.1
sqlalchemy==2.0.23
jupyter==1.0.0
ipykernel==6.26.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0